username to local storage, username validation to server, deleting ideas, preparing for deployment

+++ routes/ideas.js
@@ -45,17 +45,27 @@ router.post('/', async (req, res) => {
 //Update idea --put request
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    //get idea we're deleteing bc we need to check the username on that idea, then validate it vs the username being sent to thte routxs
+    const idea = await Idea.findById(req.params.id);
+
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+    //Username does not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to update this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -75,11 +85,21 @@ router.delete('/:id', async (req, res) => {
   //   ideas.splice(index, 1);
 
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    //get idea we're deleteing bc we need to check the username on that idea, then validate it vs the username being sent to thte rout
+    const idea = await Idea.findById(req.params.id);
+    //Match usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+    //If Usernames don't match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
 });
-http: module.exports = router;
+module.exports = router;
